refactor: extract world runtime interfaces
- The interfaces were just extracted to individual files. Almost no extra modifications performed.
