+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/BaseComponent.cs
@@ -1,69 +1,10 @@
 using DCL.Controllers;
 using DCL.Models;
-using System;
 using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace DCL.Components
 {
-    public interface IDelayedComponent : IComponent, ICleanable
-    {
-        WaitForComponentUpdate yieldInstruction { get; }
-        Coroutine routine { get; }
-        bool isRoutineRunning { get; }
-    }
-
-    public interface IMonoBehaviour
-    {
-        Transform GetTransform();
-    }
-
-    public interface IEntityComponent : IComponent, ICleanable, IMonoBehaviour
-    {
-        IDCLEntity entity { get; }
-        void Initialize(IParcelScene scene, IDCLEntity entity);
-    }
-
-    public interface ISharedComponent : IComponent, IDisposable
-    {
-        string id { get; }
-        void AttachTo(IDCLEntity entity, Type overridenAttachedType = null);
-        void DetachFrom(IDCLEntity entity, Type overridenAttachedType = null);
-        void DetachFromEveryEntity();
-        void Initialize(IParcelScene scene, string id);
-        HashSet<IDCLEntity> GetAttachedEntities();
-        void CallWhenReady(Action<ISharedComponent> callback);
-    }
-
-    public interface IComponent
-    {
-        IParcelScene scene { get; }
-        string componentName { get; }
-        void UpdateFromJSON(string json);
-        void UpdateFromModel(BaseModel model);
-        IEnumerator ApplyChanges(BaseModel model);
-        void RaiseOnAppliedChanges();
-        bool IsValid();
-        BaseModel GetModel();
-        int GetClassId();
-    }
-
-    /// <summary>
-    /// Unity is unable to yield a coroutine while is already being yielded by another one.
-    /// To fix that we wrap the routine in a CustomYieldInstruction.
-    /// </summary>
-    public class WaitForComponentUpdate : CleanableYieldInstruction
-    {
-        public IDelayedComponent component;
-
-        public WaitForComponentUpdate(IDelayedComponent component) { this.component = component; }
-
-        public override bool keepWaiting { get { return component.isRoutineRunning; } }
-
-        public override void Cleanup() { component.Cleanup(); }
-    }
-
     public abstract class BaseComponent : MonoBehaviour, IEntityComponent, IDelayedComponent, IPoolLifecycleHandler, IPoolableObjectContainer
     {
         protected ComponentUpdateHandler updateHandler;
