+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Interfaces/IComponent.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f2a79dfe718943aea8c48c4d38244ba6
+timeCreated: 1620762287
