+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Interfaces/IDelayedComponent.cs
@@ -1,0 +1,11 @@
+﻿using UnityEngine;
+
+namespace DCL.Components
+{
+    public interface IDelayedComponent : IComponent, ICleanable
+    {
+        WaitForComponentUpdate yieldInstruction { get; }
+        Coroutine routine { get; }
+        bool isRoutineRunning { get; }
+    }
+}
