+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Interfaces/IEntityComponent.cs
@@ -1,0 +1,11 @@
+﻿using DCL.Controllers;
+using DCL.Models;
+
+namespace DCL.Components
+{
+    public interface IEntityComponent : IComponent, ICleanable, IMonoBehaviour
+    {
+        IDCLEntity entity { get; }
+        void Initialize(IParcelScene scene, IDCLEntity entity);
+    }
+}
