+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Interfaces/IMonoBehaviour.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace DCL.Components
+{
+    public interface IMonoBehaviour
+    {
+        Transform GetTransform();
+    }
+}
