+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/Interfaces/IParcelScene.cs
@@ -1,0 +1,28 @@
+﻿using System.Collections.Generic;
+using DCL.Components;
+using DCL.Models;
+using UnityEngine;
+
+namespace DCL.Controllers
+{
+    public interface IParcelScene
+    {
+        Transform GetSceneTransform();
+        Dictionary<string, IDCLEntity> entities { get; }
+        Dictionary<string, ISharedComponent> disposableComponents { get; }
+        T GetSharedComponent<T>() where T : class;
+        ISharedComponent GetSharedComponent(string id);
+        event System.Action<IDCLEntity> OnEntityAdded;
+        event System.Action<IDCLEntity> OnEntityRemoved;
+        LoadParcelScenesMessage.UnityParcelScene sceneData { get; }
+        ContentProvider contentProvider { get; }
+        bool isPersistent { get; }
+        bool isTestScene { get; }
+        float loadingProgress { get; }
+        bool IsInsideSceneBoundaries(DCLCharacterPosition charPosition);
+        bool IsInsideSceneBoundaries(Bounds objectBounds);
+        bool IsInsideSceneBoundaries(Vector2Int gridPosition, float height = 0f);
+        bool IsInsideSceneBoundaries(Vector3 worldPosition, float height = 0f);
+        void CalculateSceneLoadingState();
+    }
+}
