+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/ParcelScene.cs
@@ -10,27 +10,6 @@ using UnityEngine.Assertions;
 
 namespace DCL.Controllers
 {
-    public interface IParcelScene
-    {
-        Transform GetSceneTransform();
-        Dictionary<string, IDCLEntity> entities { get; }
-        Dictionary<string, ISharedComponent> disposableComponents { get; }
-        T GetSharedComponent<T>() where T : class;
-        ISharedComponent GetSharedComponent(string id);
-        event System.Action<IDCLEntity> OnEntityAdded;
-        event System.Action<IDCLEntity> OnEntityRemoved;
-        LoadParcelScenesMessage.UnityParcelScene sceneData { get; }
-        ContentProvider contentProvider { get; }
-        bool isPersistent { get; }
-        bool isTestScene { get; }
-        float loadingProgress { get; }
-        bool IsInsideSceneBoundaries(DCLCharacterPosition charPosition);
-        bool IsInsideSceneBoundaries(Bounds objectBounds);
-        bool IsInsideSceneBoundaries(Vector2Int gridPosition, float height = 0f);
-        bool IsInsideSceneBoundaries(Vector3 worldPosition, float height = 0f);
-        void CalculateSceneLoadingState();
-    }
-
     public class ParcelScene : MonoBehaviour, IParcelScene
     {
         public static bool VERBOSE = false;
